requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
attrs==19.3.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.4.5.1
chardet==3.0.4
Django==2.1.4
django-celery==3.3.0
django-redis==4.10.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
jsonschema==3.2.0
kombu==3.0.37
mysqlclient==1.4.6
psycopg2-binary==2.8.5
pyrsistent==0.16.0
pytz==2019.3
redis==2.10.6
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
zipp==3.1.0
